chore: initial project structure
TL;DR: demos as examples, default node as an installable binary.

The "send" and "receive" (recv) demos are extremely basic, to the
point of being useless in real work. They'll only really serve, as
you can guess, as examples of how to properly use the default node.

Similarly, the node is going to be fairly simple. However, it's going
to be a much more useful simple version, since it'll be perfectly
capable of sending and receiving arbitrary data over several common
connection methods. At the same time, it'll prioritize readable code
over efficiency, to ensure it remains a good example. It's listed
as a binary so it can be installed and used as-is, if it meets your
needs.