FEAT: signatures!
You can now sign outgoing packets with a given key, and enforce that your mesher should only accept packets if they've been signed with an approved key!

The API is still a bit barebones, and will be cleaned up as I start to use it in mesher_node's development and figure out what the cleanest arrangement would be.